warning fixes

+++ src/ice-t/draw.c
@@ -447,10 +447,10 @@ void icetDrawFrame(void)
         if (zfar   >  1.0) zfar = 1.0;
 
       /* Use this information to build a containing viewport. */
-        contained_viewport[0] = left;
-        contained_viewport[1] = bottom;
-        contained_viewport[2] = right - left;
-        contained_viewport[3] = top - bottom;
+        contained_viewport[0] = (int)left;
+        contained_viewport[1] = (int)bottom;
+        contained_viewport[2] = (int)(right - left);
+        contained_viewport[3] = (int)(top - bottom);
 
       /* Now use this information to figure out which tiles need to be
          drawn. */
